backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
django==5.2.4
djangorestframework==3.16.0
idna==3.10
pip==24.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.1
drf_spectacular